FINERACT-2232: Capitalized income - Business events
