+++ fineract-provider/src/main/resources/db/changelog/tenant/changelog-tenant.xml
@@ -201,4 +201,5 @@
     <include file="parts/0180_add_version_field_to_loan_transaction.xml" relativeToChangelogFile="true" />
     <include file="parts/0181_add_capitalized_income_amortization_adjustment_transaction.xml" relativeToChangelogFile="true" />
     <include file="parts/0182_transaction_summary_with_asset_owner_report_fix_charge_reason_and_add_buyback_intermediate.xml" relativeToChangelogFile="true" />
+    <include file="parts/0183_add_LoanCapitalizedIncomeTransactionCreatedBusinessEvent.xml" relativeToChangelogFile="true" />
 </databaseChangeLog>
